refactor(lotto): change some controllers, machines to services
Change controller to service
- change LottoResultController to LottoResultService
- move package from controller to service
Change some machines to service
- change GradingMachine to LottoResultService
- move package from domain to service
